+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added optional metric to action representing the value adjusted or influenced by that action 
 
+### Changed
+
+* Moved 'Resume' and 'Pause' functionality from plugin interface to scheduler
+
 ## 0.4.0 - 2024-09-03
 
 ### Added